An Analysis of Kickstarter Campaigns
In this project, we helped Louise an up and coming playwright that wants to start a crowdfunding campaign for her play. Her budget was 10,000$. We used excel to determine whether there are specific factors that determine a plays success. We used these findings to help Louise set up her play for success.

Our findings showed that there were 525 total successful US theater Campaigns out of 912.
![US Theater Category Statistics](https://user-images.githubusercontent.com/88729583/129956545-8bcc1a2c-c5ea-4344-9dc2-0ff9d548e153.png)

In Great Britain, the most successful Kickstarters were in the 'theater' category.
![Successful Kickstarters in GB](https://user-images.githubusercontent.com/88729583/129957250-3c3371e0-864b-43c6-af51-81ebd3619f1e.png)

In terms of timing, the most successful Kickstarters were started in May. January, June, July and October all had approximately the same number of failed campaigns launched.
![Outcomes Based on Launch Date](https://user-images.githubusercontent.com/88729583/129957438-b9cc97bb-7738-40f2-af84-92f5e090f626.png)

When looking closely at the mean, median, standard deviation and interquartile range of the Goal and Pledged values we determined that the mean of each distribution is around the 3rd quartile. We also saw that the successful kickstarters had a smaller goal than the failed kickstarters (around half as small).

From the box plot analysis on GB we did, we determined see that the mean campaign goal is around 4,000.  This is outside the range of outliers for amount pledged. The range of outlies is around 3,000.

My recommendations for Louise based on these findings is that her campaign should begin in May and finish beginning of June. She should also have her goal amount be less than 10,000. I would suggest around 4,000-5,000 to guarantee a successful campaign.